+++ lib/docusign_click/api/accounts_api.rb
@@ -114,6 +114,9 @@ module DocuSign_Click
     attr_accessor :page_number
 
     # 
+    attr_accessor :shared
+
+    # 
     attr_accessor :status
 
     # 
@@ -762,19 +765,15 @@ module DocuSign_Click
       fail ArgumentError, "Missing the required parameter 'account_id' when calling AccountsApi.get_clickwrap_agreements" if account_id.nil?
       # verify the required parameter 'clickwrap_id' is set
       fail ArgumentError, "Missing the required parameter 'clickwrap_id' when calling AccountsApi.get_clickwrap_agreements" if clickwrap_id.nil?
-      # verify the required parameter 'client_user_id' is set
-      fail ArgumentError, "Missing the required parameter 'client_user_id' when calling AccountsApi.get_clickwrap_agreements" if (options.nil? or options.client_user_id.nil?)
-      # verify the required parameter 'status' is set
-      fail ArgumentError, "Missing the required parameter 'status' when calling AccountsApi.get_clickwrap_agreements" if (options.nil? or options.status.nil?)
       # resource path
       local_var_path = "/v1/accounts/{accountId}/clickwraps/{clickwrapId}/users".sub('{format}','json').sub('{' + 'accountId' + '}', account_id.to_s).sub('{' + 'clickwrapId' + '}', clickwrap_id.to_s)
 
       # query parameters
       query_params = {}
-      query_params[:'client_user_id'] = options.client_user_id
-      query_params[:'status'] = options.status
+      query_params[:'client_user_id'] = options.client_user_id if !options.client_user_id.nil?
       query_params[:'from_date'] = options.from_date if !options.from_date.nil?
       query_params[:'page_number'] = options.page_number if !options.page_number.nil?
+      query_params[:'status'] = options.status if !options.status.nil?
       query_params[:'to_date'] = options.to_date if !options.to_date.nil?
 
       # header parameters
@@ -887,14 +886,12 @@ module DocuSign_Click
       fail ArgumentError, "Missing the required parameter 'clickwrap_id' when calling AccountsApi.get_clickwrap_version_agreements" if clickwrap_id.nil?
       # verify the required parameter 'version_id' is set
       fail ArgumentError, "Missing the required parameter 'version_id' when calling AccountsApi.get_clickwrap_version_agreements" if version_id.nil?
-      # verify the required parameter 'client_user_id' is set
-      fail ArgumentError, "Missing the required parameter 'client_user_id' when calling AccountsApi.get_clickwrap_version_agreements" if (options.nil? or options.client_user_id.nil?)
       # resource path
       local_var_path = "/v1/accounts/{accountId}/clickwraps/{clickwrapId}/versions/{versionId}/users".sub('{format}','json').sub('{' + 'accountId' + '}', account_id.to_s).sub('{' + 'clickwrapId' + '}', clickwrap_id.to_s).sub('{' + 'versionId' + '}', version_id.to_s)
 
       # query parameters
       query_params = {}
-      query_params[:'client_user_id'] = options.client_user_id
+      query_params[:'client_user_id'] = options.client_user_id if !options.client_user_id.nil?
       query_params[:'from_date'] = options.from_date if !options.from_date.nil?
       query_params[:'page_number'] = options.page_number if !options.page_number.nil?
       query_params[:'status'] = options.status if !options.status.nil?
@@ -953,14 +950,12 @@ module DocuSign_Click
       fail ArgumentError, "Missing the required parameter 'clickwrap_id' when calling AccountsApi.get_clickwrap_version_agreements_by_number" if clickwrap_id.nil?
       # verify the required parameter 'version_number' is set
       fail ArgumentError, "Missing the required parameter 'version_number' when calling AccountsApi.get_clickwrap_version_agreements_by_number" if version_number.nil?
-      # verify the required parameter 'client_user_id' is set
-      fail ArgumentError, "Missing the required parameter 'client_user_id' when calling AccountsApi.get_clickwrap_version_agreements_by_number" if (options.nil? or options.client_user_id.nil?)
       # resource path
       local_var_path = "/v1/accounts/{accountId}/clickwraps/{clickwrapId}/versions/{versionNumber}/users".sub('{format}','json').sub('{' + 'accountId' + '}', account_id.to_s).sub('{' + 'clickwrapId' + '}', clickwrap_id.to_s).sub('{' + 'versionNumber' + '}', version_number.to_s)
 
       # query parameters
       query_params = {}
-      query_params[:'client_user_id'] = options.client_user_id
+      query_params[:'client_user_id'] = options.client_user_id if !options.client_user_id.nil?
       query_params[:'from_date'] = options.from_date if !options.from_date.nil?
       query_params[:'page_number'] = options.page_number if !options.page_number.nil?
       query_params[:'status'] = options.status if !options.status.nil?
@@ -1129,6 +1124,7 @@ module DocuSign_Click
       query_params[:'from_date'] = options.from_date if !options.from_date.nil?
       query_params[:'ownerUserId'] = options.owner_user_id if !options.owner_user_id.nil?
       query_params[:'page_number'] = options.page_number if !options.page_number.nil?
+      query_params[:'shared'] = options.shared if !options.shared.nil?
       query_params[:'status'] = options.status if !options.status.nil?
       query_params[:'to_date'] = options.to_date if !options.to_date.nil?