+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # DocuSign Click Ruby Client Changelog
 See [DocuSign Support Center](https://support.docusign.com/en/releasenotes/) for Product Release Notes.
 
+## [v1.0.0] - DocuSign Click API v1 - 02/11/2021
+### Changed
+- First GA version of Click API, supports DocuSign Click v1-20.4.02.
+
 ## [v1.0.0.beta] - Click API v1-20.3.02 - 2020-12-17
 ### Changed
 - First BETA version of Click API